refactor(prep_naming): #556 — consolidate abbrev_title() duplication (M3+)
abbrev_title() was duplicated between findajob.prep.orchestrator (typed)
and scripts/rename_folders.py (untyped). Consolidated to
findajob.prep_naming — same domain as safe_filename_part (both build
folder-safe filename components). Both call sites import from there.

No re-export theater per the M3 cleanup discipline (#544): orchestrator
no longer exposes abbrev_title at its module level. test_prep_import_safety
updated to assert the canonical import location.

First M3+ punch-list item from the closed M3 epic (#537).

Verification:
- uv run pytest -q → 1971 passed, 1 skipped (pre-existing)
- uv run ruff check src/ scripts/ tests/ → All checks passed
- uv run ruff format --check → 288 files formatted
- uv run mypy src/findajob/ → no issues found in 113 source files

Closes #556.
